Задача MVVM
