+++ app/src/main/java/ru/netology/nmedia/repository/PostRepository.kt
@@ -1,0 +1,11 @@
+package ru.netology.nmedia.repository
+
+import androidx.lifecycle.LiveData
+import ru.netology.nmedia.dto.Post
+
+interface PostRepository {
+    fun get(): LiveData<Post>
+    fun like()
+    fun share()
+    fun views()
+}
